requirements.txt
asgiref==3.8.1
Django==5.1.1
django-phonenumber-field==8.0.0
phonenumberslite==8.13.45
sqlparse==0.5.1
